refactor: 패키지 이름을 util로 변경합니다.
-구현한 클래스들의 메서드가 static을 사용하고, 상태를 가지지 않아 service에서 util로 이름을 변경합니다.
-관련된 메서드 import 문의 패키지 명을 변경합니다.
